test(logic): split resolveTurnForGame part1 tests for #2288
Replace the 886-line turn_resolver_resolve_turn_for_game_part1_test.dart
with three segment files (movement/economy, tileMap/production, combat
scenarios) so each stays under the 400-line logic test incremental gate.
Test bodies and assertions unchanged.

Refs #2288
